requirements.txt
asgiref==3.10.0
Django==5.2.8
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.3
python-dotenv==0.21.0
